+++ src/collective/cover/widgets/textlinessortable.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from collective.cover.widgets.interfaces import ITextLinesSortableWidget
+from plone import api
 from plone.app.uuid.utils import uuidToObject
 from z3c.form import interfaces
 from z3c.form import widget
@@ -55,6 +56,44 @@ class TextLinesSortableWidget(textlines.TextLinesWidget):
         except:
             return None
 
+    def get_custom_title(self, uuid):
+        """ Returns the custom Title assigned to a specific item
+
+        :param uuid: [required] The object's UUID
+        :type uuid: string
+        :returns: The custom Title
+        """
+        # Try to get custom title
+        title = u''
+        uuids = self.context['uuids']
+        values = [uuids[i] for i in uuids if i == uuid]
+        if values:
+            title = values[0].get('custom_title', u'')
+        if title:
+            return title
+        # If didn't find, get object title
+        obj = uuidToObject(uuid)
+        return obj.Title()
+
+    def get_custom_description(self, uuid):
+        """ Returns the custom Description assigned to a specific item
+
+        :param uuid: [required] The object's UUID
+        :type uuid: string
+        :returns: The custom Description
+        """
+        # Try to get custom description
+        description = u''
+        uuids = self.context['uuids']
+        values = [uuids[i] for i in uuids if i == uuid]
+        if values:
+            description = values[0].get('custom_description', u'')
+        if description:
+            return description
+        # If didn't find, get object description
+        obj = uuidToObject(uuid)
+        return obj.Description()
+
     def get_custom_url(self, uuid):
         """ Returns the custom URL assigned to a specific item
 
@@ -62,11 +101,22 @@ class TextLinesSortableWidget(textlines.TextLinesWidget):
         :type uuid: string
         :returns: The custom URL
         """
+        # Try to get custom url
         url = u''
         uuids = self.context['uuids']
-        if uuid in uuids:
-            values = uuids.get(uuid)
-            url = values.get('custom_url', u'')
+        values = [uuids[i] for i in uuids if i == uuid]
+        if values:
+            url = values[0].get('custom_url', u'')
+        if url:
+            return url
+        # If didn't find, get object url
+        obj = uuidToObject(uuid)
+        portal_properties = api.portal.get_tool(name='portal_properties')
+        use_view_action = portal_properties.site_properties.getProperty(
+            'typesUseViewActionInListings', ())
+        url = obj.absolute_url()
+        if obj.portal_type in use_view_action:
+            url = url + '/view'
         return url
 
     def extract(self):
@@ -78,14 +128,21 @@ class TextLinesSortableWidget(textlines.TextLinesWidget):
         uuids = [i for i in values if i]
         results = dict()
         for index, uuid in enumerate(uuids):
-            if uuid:
-                custom_url = self.request.get(
-                    '%s.custom_url.%s' % (self.name, uuid), ""
-                )
-                results[uuid] = {
-                    u'order': unicode(index),
-                    u'custom_url': unicode(custom_url)
-                }
+            custom_title = self.request.get(
+                '{0}.custom_title.{1}'.format(self.name, uuid), ''
+            )
+            custom_description = self.request.get(
+                '{0}.custom_description.{1}'.format(self.name, uuid), ''
+            )
+            custom_url = self.request.get(
+                '{0}.custom_url.{1}'.format(self.name, uuid), ''
+            )
+            results[uuid] = {
+                u'order': unicode(index),
+                u'custom_title': unicode(custom_title),
+                u'custom_description': unicode(custom_description),
+                u'custom_url': unicode(custom_url)
+            }
         return results